Added MaxsSearch() function which returns a vector of nodes where the key is equal to the search key. Called it MaxsSearch because I wasn’t sure what complete search is supposed to do so I left that function alone in case it is supposed to have another purpose
Edited menu items D and E to work with MaxsSearch

Overloaded the >> operator for GeneralData to print out all non-empty fields with labels. Also added a printData() function to Node.h that uses the overloaded operator

Updated menu items D and E to allow user to edit fields

Added a single line to ReadPictureCSVFile() that erases the first character of the film name because those fuckers put a space in the .csv in front of every film name and it broke my search function and it took me like a goddamn hour to figure it out

Cleaned up / improved the functionality (I think) of menu options A and B by replacing cin’s with getline’s and reordering some things

Also, I added functions.cpp and functions.hpp, They're empty right now, but I think at some point we should move the menu to a function and compartmentalize the various menu features, as some of them are very long and likely reusable. I think it would improve our grade.